projet 14 GOFFINET-EL MOKHTARI-BERRICHI/requirements.txt
jsonschema>=4.19.0
python-dotenv>=1.0.0
reportlab>=4.0.0
PyYAML>=6.0
mypy>=1.9.0
pytest>=7.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pdfkit>=0.6.4
openai>=1.0.0
semantic-kernel>=0.9.0